fix: keep tracker logs within print frontend

+++ ArmorTracker.hpp
@@ -1150,8 +1150,8 @@ std::optional<ArmorDetectorResult> ArmorTracker<CameraInfoV>::SelectSingleArmorO
       if (suspicious_pose_jump)
       {
         XR_LOG_DEBUG(
-            "SingleArmor reject pose jump: idx=%zu track=%d yaw_jump=%.3f center_diff=%.1f area_log=%.3f prev=(%.3f,%.3f,%.3f) now=(%.3f,%.3f,%.3f)",
-            armor_index, detection_track_id, yaw_jump,
+            "SingleArmor reject pose jump: idx=%u track=%d yaw_jump=%.3f center_diff=%.1f area_log=%.3f prev=(%.3f,%.3f,%.3f) now=(%.3f,%.3f,%.3f)",
+            static_cast<unsigned>(armor_index), detection_track_id, yaw_jump,
             static_cast<double>(center_diff), static_cast<double>(area_log), rt_.tracked_armor.pose.translation.x(),
             rt_.tracked_armor.pose.translation.y(), rt_.tracked_armor.pose.translation.z(),
             armor.pose.translation.x(), armor.pose.translation.y(),
